build(style): replace to tailwind
