+++ postcss.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  plugins: ['tailwindcss']
+}
