--- src/context/ThemeContext/Theme.ts
@@ -1,3 +1,0 @@
-export type Theme = 'light' | 'dark'
-
-export default Theme
