--- src/context/ThemeContext/ThemeContext.ts
@@ -1,14 +1,0 @@
-import { createContext } from 'react'
-
-import Theme from './Theme'
-
-type Value = {
-  theme: Theme
-  toggleTheme?: () => void
-}
-
-const ThemeContext = createContext<Value>({
-  theme: 'light'
-})
-
-export default ThemeContext
