--- src/context/ThemeContext/index.ts
@@ -1,4 +1,0 @@
-export type { default as Theme } from './Theme'
-export { default as ThemeProvider } from './ThemeProvider'
-export { default as useTheme } from './useTheme'
-export { default } from './ThemeContext'
