--- src/context/ThemeContext/useTheme.ts
@@ -1,15 +1,0 @@
-import { useContext } from 'react'
-
-import type Theme from './Theme'
-import ThemeContext from './ThemeContext'
-
-// eslint-disable-next-line @typescript-eslint/no-empty-function
-function noop(): void {}
-
-function useTheme(): [Theme, () => void] {
-  const { theme, toggleTheme = noop } = useContext(ThemeContext)
-
-  return [theme, toggleTheme]
-}
-
-export default useTheme
