+++ src/pages/404.mdx
@@ -1,4 +1,4 @@
-import Markdown from '@/components/templates/Markdown'
+import Markdown from '@/components/Markdown'
 
 export const title = '404 Not Found!'
