+++ src/pages/about.mdx
@@ -1,4 +1,4 @@
-import Markdown from '@/components/templates/Markdown'
+import Markdown from '@/components/Markdown'
 
 export const title = 'SHINJU DATEとは?'
