+++ src/pages/terms.mdx
@@ -1,4 +1,4 @@
-import Markdown from '@/components/templates/Markdown'
+import Markdown from '@/components/Markdown'
 
 export const title = '利用規約'
