+++ tailwind.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  future: {
+    removeDeprecatedGapUtilities: true
+  },
+  plugins: [],
+  purge: ['./src/**/*.{mdx,ts,tsx}'],
+  theme: {
+    extend: {}
+  },
+  variants: {}
+}
